+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/dialog/LoginPanel.java
@@ -149,9 +149,9 @@ final class LoginPanel extends JPanel {
 	private JPanel createCredentialsPanel() {
 		return PanelBuilder.builder()
 						.flexibleGridLayout(2, 2)
-						.add(new JLabel(Messages.username(), SwingConstants.RIGHT))
+						.add(new JLabel(Messages.username(), SwingConstants.LEADING))
 						.add(usernameField)
-						.add(new JLabel(Messages.password(), SwingConstants.RIGHT))
+						.add(new JLabel(Messages.password(), SwingConstants.LEADING))
 						.add(createPasswordProgressPanel())
 						.build();
 	}